// model/aws/wellarchitected/extensions/models/_lib/aws.ts
export interface AwsCredentials {
  accessKeyId?: string;
  secretAccessKey?: string;
  sessionToken?: string;
  region?: string;
}